backend/apps/api/viewsets/appointments.py
from rest_framework.viewsets import ModelViewSet
from django_filters.rest_framework import DjangoFilterBackend

from appointments.models import Appointment
from api.serializers.appointments import AppointmentSerializer


class AppointmentsViewset(ModelViewSet):
    queryset = Appointment.objects.all()
    serializer_class = AppointmentSerializer
    filter_backends = [DjangoFilterBackend]
    filterset_fields = ["office"]
