backend/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
baron==0.10.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-email-url==1.0.6
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cache-url==3.4.4
django-cors-headers==3.14.0
django-environs==0.1.1
django-filter==23.1
django-iranian-cities @ git+https://github.com/sageteam-org/django-iranian-cities.git@b4d5f63a19acb8555e54b9116c54cb3e1723c7e9
django-jalali==6.0.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
gunicorn==20.1.0
idna==3.4
jdatetime==4.1.1
Markdown==3.4.3
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
persiantools==3.0.1
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redbaron==0.9.2
requests==2.29.0
requests-oauthlib==1.3.1
rply==0.7.8
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
